Re organize CLI tools into separate packages
