+++ cli/admin/README.md
@@ -1,0 +1,1 @@
+# Jumpstarter Admin CLI
