+++ cli/admin/jumpstarter_admin_cli/delete.py
@@ -2,13 +2,7 @@ import logging
 from typing import Optional
 
 import asyncclick as click
-from kubernetes_asyncio.client.exceptions import ApiException
-from kubernetes_asyncio.config.config_exception import ConfigException
-
-from jumpstarter.config import ClientConfigV1Alpha1, ExporterConfigV1Alpha1, UserConfigV1Alpha1
-from jumpstarter.k8s import ClientsV1Alpha1Api, ExportersV1Alpha1Api
-
-from .util import (
+from jumpstarter_cli_common import (
     AliasedGroup,
     handle_k8s_api_exception,
     handle_k8s_config_exception,
@@ -17,6 +11,11 @@ from .util import (
     opt_log_level,
     opt_namespace,
 )
+from kubernetes_asyncio.client.exceptions import ApiException
+from kubernetes_asyncio.config.config_exception import ConfigException
+
+from jumpstarter.config import ClientConfigV1Alpha1, ExporterConfigV1Alpha1, UserConfigV1Alpha1
+from jumpstarter.k8s import ClientsV1Alpha1Api, ExportersV1Alpha1Api
 
 
 @click.group(cls=AliasedGroup)
