+++ cli/admin/jumpstarter_admin_cli/delete_test.py
