+++ cli/admin/jumpstarter_admin_cli/import_res.py
@@ -1,6 +1,13 @@
 from typing import Optional
 
 import asyncclick as click
+from jumpstarter_cli_common import (
+    handle_k8s_api_exception,
+    handle_k8s_config_exception,
+    opt_context,
+    opt_kubeconfig,
+    opt_namespace,
+)
 from kubernetes_asyncio.client.exceptions import ApiException
 from kubernetes_asyncio.config.config_exception import ConfigException
 
@@ -10,8 +17,6 @@ from jumpstarter.config import (
 )
 from jumpstarter.config.exporter import ExporterConfigV1Alpha1
 from jumpstarter.k8s import ClientsV1Alpha1Api, ExportersV1Alpha1Api
-
-from .util import handle_k8s_api_exception, handle_k8s_config_exception, opt_context, opt_kubeconfig, opt_namespace
 
 
 @click.group("import")
