+++ cli/admin/pyproject.toml
@@ -1,0 +1,48 @@
+[project]
+name = "jumpstarter-admin-cli"
+dynamic = ["version", "urls"]
+description = ""
+authors = [
+  { name = "Nick Cao", email = "ncao@redhat.com" },
+  { name = "Miguel Angel Ajo Pelayo", email = "majopela@redhat.com" },
+  { name = "Kirk Brauer", email = "kbrauer@hatci.com" },
+]
+readme = "README.md"
+license = { text = "Apache-2.0" }
+requires-python = ">=3.11"
+dependencies = [
+    "jumpstarter",
+    "jumpstarter-cli-common",
+    "kubernetes>=31.0.0",
+    "kubernetes-asyncio>=31.1.0",
+    "asyncclick>=8.1.7.2",
+]
+
+[dependency-groups]
+dev = [
+    "pytest>=8.3.2",
+    "pytest-anyio>=0.0.0",
+    "pytest-asyncio>=0.0.0",
+    "pytest-cov>=5.0.0",
+]
+
+[tool.uv.sources]
+jumpstarter-cli-common = { workspace = true }
+
+[project.scripts]
+jmp-admin = "jumpstarter_admin_cli:admin"
+
+[tool.hatch.build.targets.wheel]
+packages = ["jumpstarter_admin_cli"]
+
+[tool.hatch.metadata.hooks.vcs.urls]
+Homepage = "https://jumpstarter.dev"
+source_archive = "https://github.com/jumpstarter-dev/repo/archive/{commit_hash}.zip"
+
+[tool.hatch.version]
+source = "vcs"
+raw-options = { 'root' = '../../'}
+
+[build-system]
+requires = ["hatchling", "hatch-vcs"]
+build-backend = "hatchling.build"
