+++ cli/client/README.md
@@ -1,0 +1,1 @@
+# Jumpstarter Client CLI
