+++ cli/common/jumpstarter_cli_common/__init__.py
@@ -3,6 +3,7 @@ from .k8s import handle_k8s_api_exception, handle_k8s_config_exception
 from .opt import opt_context, opt_kubeconfig, opt_log_level, opt_namespace
 from .table import make_table
 from .time import time_since
+from .version import get_client_version, version
 
 __all__ = [
     "AliasedGroup",
@@ -13,5 +14,7 @@ __all__ = [
     "opt_namespace",
     "time_since",
     "handle_k8s_api_exception",
-    "handle_k8s_config_exception"
+    "handle_k8s_config_exception",
+    "version",
+    "get_client_version"
 ]
