+++ cli/common/jumpstarter_cli_common/alias.py
@@ -7,11 +7,19 @@ class AliasedGroup(click.Group):
     common_aliases: dict[str, list[str]] = {
         "remove": ["rm"],
         "list": ["ls"],
-        "create": ["cr"],
+        "create": ["c"],
+        "import": ["i"],
+        "get": ["g"],
+        "admin": ["a"],
+        "create-config": ["cc"],
+        "delete-config": ["dc"],
+        "edit-config": ["ec"],
+        "list-configs": ["lc"],
+        "use-config": ["uc"],
         "move": ["mv"],
         "config": ["conf"],
-        "delete": ["del"],
-        "shell": ["sh"],
+        "delete": ["del", "d"],
+        "shell": ["sh", "s"],
         "exporter": ["exporters", "e"],
         "client": ["clients", "c"],
         "lease": ["leases", "l"],
