+++ cli/common/jumpstarter_cli_common/table.py
