+++ cli/jmp/README.md
@@ -1,0 +1,1 @@
+# The Jumpstarter CLI
